Relocate artifact metadata provider to repos directory
