+++ Porytiles2/include/porytiles2/app/use_cases/compile_primary_tileset.hpp
@@ -5,7 +5,6 @@
 
 #include "porytiles2/domain/repos/tileset_repo.hpp"
 #include "porytiles2/domain/services/primary_tileset_compiler.hpp"
-#include "porytiles2/domain/services/tileset_artifact_metadata_provider.hpp"
 #include "porytiles2/templates/result.hpp"
 
 namespace porytiles2 {
