+++ Porytiles2/lib/domain/repos/tileset_artifact_metadata_provider.cpp
@@ -1,4 +1,4 @@
-#include "porytiles2/domain/services/tileset_artifact_metadata_provider.hpp"
+#include "porytiles2/domain/repos/tileset_artifact_metadata_provider.hpp"
 
 #include <string>
 #include <unordered_map>
@@ -8,7 +8,7 @@ namespace porytiles2 {
 
 std::vector<std::string>
 TilesetArtifactMetadataProvider::find_unsynced_artifacts(const std::string &tileset_name,
-                                                  const std::vector<std::string> &artifact_keys) const {
+                                                         const std::vector<std::string> &artifact_keys) const {
     const auto checksums = compute_artifact_checksums(tileset_name);
     const auto cached_checksums = load_cached_checksums(tileset_name);
 
@@ -25,7 +25,7 @@ TilesetArtifactMetadataProvider::find_unsynced_artifacts(const std::string &tile
 }
 
 bool TilesetArtifactMetadataProvider::all_checksums_match(const std::string &tileset_name,
-                                                   const std::vector<std::string> &artifact_keys) const {
+                                                          const std::vector<std::string> &artifact_keys) const {
     return find_unsynced_artifacts(tileset_name, artifact_keys).empty();
 }
